Deadline Commit
Unfortunately I was unable to complete the assignment to the best of my ability before the deadline. For what it's worth, I did consult with my instructors and even they had a challenge helping me with my logic. I am confident I could complete it, but I regret I have to submit incomplete.